Add StreamBuilder into its own file to ensure private fields are respected
