+++ banyan/src/tree.rs
@@ -7,81 +7,12 @@ use crate::{
     store::BlockWriter,
     util::BoxedIter,
 };
-use crate::{query::Query, store::ReadOnlyStore, util::IterExt};
+use crate::{query::Query, store::ReadOnlyStore, util::IterExt, StreamBuilder, StreamBuilderState};
 use anyhow::Result;
 use futures::prelude::*;
 use libipld::cbor::DagCbor;
-use std::{collections::BTreeMap, fmt, fmt::Debug, iter, sync::Arc, usize};
+use std::{collections::BTreeMap, fmt::Debug, iter, sync::Arc, usize};
 use tracing::*;
-
-/// A thing that hands out unique offsets. Parts of StreamBuilderState
-#[derive(Debug, Clone)]
-pub struct Offset {
-    value: u64,
-}
-
-impl Offset {
-    pub fn new(value: u64) -> Self {
-        Self { value }
-    }
-
-    pub fn reserve(&mut self, n: usize) -> u64 {
-        let result = self.value;
-        self.value = self
-            .value
-            .checked_add(n as u64)
-            .expect("ran out of offsets");
-        result
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct StreamBuilderState {
-    /// the secrets used for building the trees of the stream
-    secrets: Secrets,
-    /// config that determines the branch size etc.
-    pub(crate) config: Config,
-    /// current offset
-    pub(crate) offset: Offset,
-}
-
-impl StreamBuilderState {
-    pub fn new(offset: u64, secrets: Secrets, config: Config) -> Self {
-        Self {
-            offset: Offset::new(offset),
-            secrets,
-            config,
-        }
-    }
-
-    pub fn zstd_level(&self) -> i32 {
-        self.config().zstd_level
-    }
-
-    pub fn config(&self) -> &Config {
-        &self.config
-    }
-
-    pub fn secrets(&self) -> &Secrets {
-        &self.secrets
-    }
-
-    pub fn value_key(&self) -> &chacha20::Key {
-        self.secrets.value_key()
-    }
-
-    pub fn index_key(&self) -> &chacha20::Key {
-        self.secrets.index_key()
-    }
-}
-
-/// A tree. This is mostly an user friendly handle.
-///
-/// Most of the logic except for handling the empty case is implemented in the forest
-pub struct StreamBuilder<T: TreeTypes> {
-    root: Option<Arc<Index<T>>>,
-    state: StreamBuilderState,
-}
 
 #[derive(Debug, Clone)]
 pub struct Tree<T: TreeTypes> {
@@ -100,48 +31,16 @@ impl<T: TreeTypes> Tree<T> {
     }
 
     pub(crate) fn into_inner(self) -> Option<(Arc<Index<T>>, Secrets)> {
-        let secrets = self.secrets.clone();
+        let secrets = self.secrets().clone();
         self.root.map(|x| (x, secrets))
     }
 
     pub(crate) fn as_index_ref(&self) -> Option<&Index<T>> {
         self.root.as_ref().map(|arc| arc.as_ref())
     }
-}
 
-impl<T: TreeTypes> fmt::Debug for StreamBuilder<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.root {
-            Some(root) => f
-                .debug_struct("Tree")
-                .field("count", &self.count())
-                .field("key_bytes", &root.key_bytes())
-                .field("value_bytes", &root.value_bytes())
-                .field("link", &root.link())
-                .finish(),
-            None => f
-                .debug_struct("Tree")
-                .field("count", &self.count())
-                .finish(),
-        }
-    }
-}
-
-impl<T: TreeTypes> fmt::Display for StreamBuilder<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.root {
-            Some(root) => write!(f, "{:?}", root.link(),),
-            None => write!(f, "empty tree"),
-        }
-    }
-}
-
-impl<T: TreeTypes> Clone for StreamBuilder<T> {
-    fn clone(&self) -> Self {
-        Self {
-            root: self.root.clone(),
-            state: self.state.clone(),
-        }
+    pub fn secrets(&self) -> &Secrets {
+        &self.secrets
     }
 }
 
@@ -179,7 +78,7 @@ impl<
     /// dumps the tree structure
     pub fn dump(&self, tree: &Tree<T>) -> Result<()> {
         match &tree.root {
-            Some(index) => self.dump0(&tree.secrets, index, ""),
+            Some(index) => self.dump0(tree.secrets(), index, ""),
             None => Ok(()),
         }
     }
@@ -191,7 +90,7 @@ impl<
         f: impl Fn((usize, &NodeInfo<T>)) -> S + Clone,
     ) -> Result<(GraphEdges, GraphNodes<S>)> {
         match &tree.root {
-            Some(index) => self.dump_graph0(&tree.secrets, None, 0, index, f),
+            Some(index) => self.dump_graph0(tree.secrets(), None, 0, index, f),
             None => anyhow::bail!("Tree must not be empty"),
         }
     }
@@ -244,7 +143,7 @@ impl<
 
     pub(crate) fn dump_graph0<S>(
         &self,
-        stream: &Secrets,
+        secrets: &Secrets,
         parent_id: Option<usize>,
         next_id: usize,
         index: &Index<T>,
@@ -253,7 +152,7 @@ impl<
         let mut edges = vec![];
         let mut nodes: BTreeMap<usize, S> = Default::default();
 
-        let node = self.load_node(stream, index)?;
+        let node = self.load_node(secrets, index)?;
         if let Some(p) = parent_id {
             edges.push((p, next_id));
         }
@@ -262,7 +161,7 @@ impl<
             let mut cur = next_id;
             for x in branch.children.iter() {
                 let (mut e, mut n) =
-                    self.dump_graph0(stream, Some(next_id), cur + 1, &x, f.clone())?;
+                    self.dump_graph0(secrets, Some(next_id), cur + 1, &x, f.clone())?;
                 cur += n.len();
                 edges.append(&mut e);
                 nodes.append(&mut n);
@@ -274,8 +173,8 @@ impl<
 
     /// sealed roots of the tree
     pub fn roots(&self, tree: &StreamBuilder<T>) -> Result<Vec<Index<T>>> {
-        match &tree.root {
-            Some(index) => self.roots_impl(tree.state.secrets(), index),
+        match tree.index() {
+            Some(index) => self.roots_impl(tree.state().secrets(), index),
             None => Ok(Vec::new()),
         }
     }
@@ -283,9 +182,9 @@ impl<
     /// leftmost branches of the tree as separate trees
     pub fn left_roots(&self, tree: &Tree<T>) -> Result<Vec<Tree<T>>> {
         Ok(if let Some(index) = tree.as_index_ref() {
-            self.left_roots0(&tree.secrets, index)?
+            self.left_roots0(tree.secrets(), index)?
                 .into_iter()
-                .map(|x| Tree::new(Some(Arc::new(x)), tree.secrets.clone(), u64::max_value()))
+                .map(|x| Tree::new(Some(Arc::new(x)), tree.secrets().clone(), u64::max_value()))
                 .collect()
         } else {
             Vec::new()
@@ -309,11 +208,11 @@ impl<
 
     pub fn check_invariants(&self, tree: &StreamBuilder<T>) -> Result<Vec<String>> {
         let mut msgs = Vec::new();
-        if let Some(root) = &tree.root {
+        if let Some(root) = tree.index() {
             let mut level = i32::max_value();
             self.check_invariants0(
-                tree.state.secrets(),
-                tree.state.config(),
+                tree.state().secrets(),
+                tree.state().config(),
                 root,
                 &mut level,
                 &mut msgs,
@@ -324,7 +223,7 @@ impl<
 
     pub fn is_packed(&self, tree: &Tree<T>) -> Result<bool> {
         if let Some(root) = &tree.root {
-            self.is_packed0(&tree.secrets, &root)
+            self.is_packed0(tree.secrets(), &root)
         } else {
             Ok(true)
         }
@@ -349,7 +248,7 @@ impl<
     ) -> impl Stream<Item = Result<(u64, T::Key, V)>> + 'static {
         match &tree.root {
             Some(index) => self
-                .stream_filtered0(tree.secrets.clone(), query, index.clone())
+                .stream_filtered0(tree.secrets().clone(), query, index.clone())
                 .left_stream(),
             None => stream::empty().right_stream(),
         }
@@ -364,7 +263,7 @@ impl<
     ) -> impl Iterator<Item = Result<Arc<Index<T>>>> + 'static {
         match &tree.root {
             Some(index) => self
-                .index_iter0(tree.secrets.clone(), query, index.clone())
+                .index_iter0(tree.secrets().clone(), query, index.clone())
                 .boxed()
                 .left_iter(),
             None => iter::empty().right_iter(),
@@ -380,7 +279,7 @@ impl<
     ) -> impl Iterator<Item = Result<Arc<Index<T>>>> + 'static {
         match &tree.root {
             Some(index) => self
-                .index_iter_rev0(tree.secrets.clone(), query, index.clone())
+                .index_iter_rev0(tree.secrets().clone(), query, index.clone())
                 .boxed()
                 .left_iter(),
             None => iter::empty().right_iter(),
@@ -394,7 +293,7 @@ impl<
     ) -> impl Iterator<Item = Result<(u64, T::Key, V)>> + 'static {
         match &tree.root {
             Some(index) => self
-                .iter_filtered0(tree.secrets.clone(), query, index.clone())
+                .iter_filtered0(tree.secrets().clone(), query, index.clone())
                 .boxed()
                 .left_iter(),
             None => iter::empty().right_iter(),
@@ -408,7 +307,7 @@ impl<
     ) -> impl Iterator<Item = Result<(u64, T::Key, V)>> + 'static {
         match &tree.root {
             Some(index) => self
-                .iter_filtered_reverse0(tree.secrets.clone(), query, index.clone())
+                .iter_filtered_reverse0(tree.secrets().clone(), query, index.clone())
                 .boxed()
                 .left_iter(),
             None => iter::empty().right_iter(),
@@ -421,7 +320,11 @@ impl<
     ) -> impl Iterator<Item = Result<(u64, T::Key, V)>> + 'static {
         match &tree.root {
             Some(index) => self
-                .iter_filtered0(tree.secrets.clone(), crate::query::AllQuery, index.clone())
+                .iter_filtered0(
+                    tree.secrets().clone(),
+                    crate::query::AllQuery,
+                    index.clone(),
+                )
                 .left_iter(),
             None => iter::empty().right_iter(),
         }
@@ -440,7 +343,7 @@ impl<
     {
         match &tree.root {
             Some(index) => self
-                .traverse0(tree.secrets.clone(), query, index.clone(), mk_extra)
+                .traverse0(tree.secrets().clone(), query, index.clone(), mk_extra)
                 .left_iter(),
             None => iter::empty().right_iter(),
         }
@@ -459,7 +362,7 @@ impl<
     {
         match &tree.root {
             Some(index) => self
-                .traverse_rev0(tree.secrets.clone(), query, index.clone(), mk_extra)
+                .traverse_rev0(tree.secrets().clone(), query, index.clone(), mk_extra)
                 .left_iter(),
             None => iter::empty().right_iter(),
         }
@@ -478,7 +381,7 @@ impl<
     {
         match &tree.root {
             Some(index) => self
-                .stream_filtered_chunked0(tree.secrets.clone(), query, index.clone(), mk_extra)
+                .stream_filtered_chunked0(tree.secrets().clone(), query, index.clone(), mk_extra)
                 .left_stream(),
             None => stream::empty().right_stream(),
         }
@@ -497,7 +400,7 @@ impl<
     {
         match &tree.root {
             Some(index) => self
-                .stream_filtered_chunked_reverse0(&tree.secrets, query, index.clone(), mk_extra)
+                .stream_filtered_chunked_reverse0(tree.secrets(), query, index.clone(), mk_extra)
                 .left_stream(),
             None => stream::empty().right_stream(),
         }
@@ -510,7 +413,7 @@ impl<
     /// not be read.
     pub fn get(&self, tree: &Tree<T>, offset: u64) -> Result<Option<(T::Key, V)>> {
         Ok(match &tree.root {
-            Some(index) => self.get0(&tree.secrets, index, offset)?,
+            Some(index) => self.get0(tree.secrets(), index, offset)?,
             None => None,
         })
     }
@@ -526,7 +429,7 @@ impl<
     pub fn collect_from(&self, tree: &Tree<T>, offset: u64) -> Result<Vec<Option<(T::Key, V)>>> {
         let mut res = Vec::new();
         if let Some(index) = &tree.root {
-            self.collect0(&tree.secrets, index, offset, &mut res)?;
+            self.collect0(&tree.secrets(), index, offset, &mut res)?;
         }
         Ok(res)
     }
@@ -539,7 +442,7 @@ impl<
         W: BlockWriter<T::Link> + 'static,
     > Transaction<T, V, R, W>
 {
-    pub fn tree_from_roots(
+    pub(crate) fn tree_from_roots(
         &self,
         mut roots: Vec<Index<T>>,
         stream: &mut StreamBuilderState,
@@ -561,7 +464,7 @@ impl<
     /// ![packing illustration](https://ipfs.io/ipfs/QmaEDTjHSdCKyGQ3cFMCf73kE67NvffLA5agquLW5qSEVn/packing.jpg)
     pub fn pack(&self, tree: &StreamBuilder<T>) -> Result<StreamBuilder<T>> {
         let roots = self.roots(tree)?;
-        let mut state = tree.state.clone();
+        let mut state = tree.state().clone();
         let filled = self.tree_from_roots(roots, &mut state)?;
         let remainder: Vec<_> = self
             .collect_from(&tree.snapshot(), filled.count())?
@@ -595,7 +498,7 @@ impl<
             // nothing to do
             return Ok(tree.clone());
         }
-        let mut state = tree.state.clone();
+        let mut state = tree.state().clone();
         let index = self.extend_above(
             tree.as_index_ref(),
             u32::max_value(),
@@ -619,7 +522,7 @@ impl<
         I: IntoIterator<Item = (T::Key, V)>,
         I::IntoIter: Send,
     {
-        let mut state = tree.state.clone();
+        let mut state = tree.state().clone();
         let index = self.extend_unpacked0(tree.as_index_ref(), from, &mut state)?;
         Ok(StreamBuilder::new_from_index(index, state))
     }
@@ -639,9 +542,9 @@ impl<
         tree: &StreamBuilder<T>,
         query: &'a Q,
     ) -> Result<StreamBuilder<T>> {
-        Ok(if let Some(index) = &tree.root {
+        Ok(if let Some(index) = tree.index() {
             let mut level: i32 = i32::max_value();
-            let mut state = tree.state.clone();
+            let mut state = tree.state().clone();
             let res = self.retain0(0, query, index, &mut level, &mut state)?;
             StreamBuilder::new_from_index(Some(res), state)
         } else {
@@ -658,72 +561,14 @@ impl<
     /// especially if there are repaired blocks in the non-packed part.
     pub fn repair(&self, tree: &StreamBuilder<T>) -> Result<(StreamBuilder<T>, Vec<String>)> {
         let mut report = Vec::new();
-        Ok(if let Some(index) = &tree.root {
+        Ok(if let Some(index) = tree.index() {
             let mut level: i32 = i32::max_value();
-            let mut state = tree.state.clone();
+            let mut state = tree.state().clone();
             let repaired = self.repair0(index, &mut report, &mut level, &mut state)?;
             (StreamBuilder::new_from_index(Some(repaired), state), report)
         } else {
             (tree.clone(), report)
         })
-    }
-}
-
-impl<T: TreeTypes> StreamBuilder<T> {
-    pub fn state(&self) -> StreamBuilderState {
-        self.state.clone()
-    }
-
-    pub fn new(config: Config, secrets: Secrets) -> Self {
-        let state = StreamBuilderState::new(0, secrets, config);
-        Self::new_from_index(None, state)
-    }
-
-    pub(crate) fn new_from_index(root: Option<Index<T>>, state: StreamBuilderState) -> Self {
-        Self {
-            root: root.map(Arc::new),
-            state,
-        }
-    }
-
-    pub fn snapshot(&self) -> Tree<T> {
-        Tree::new(
-            self.root.clone(),
-            self.state.secrets().clone(),
-            self.state.offset.value,
-        )
-    }
-
-    pub fn link(&self) -> Option<T::Link> {
-        self.root.as_ref().and_then(|r| *r.link())
-    }
-
-    pub fn as_index_ref(&self) -> Option<&Index<T>> {
-        self.root.as_ref().map(|arc| arc.as_ref())
-    }
-
-    pub fn level(&self) -> i32 {
-        self.root.as_ref().map(|x| x.level() as i32).unwrap_or(-1)
-    }
-
-    pub fn debug() -> Self {
-        let state = StreamBuilderState::new(0, Secrets::default(), Config::debug());
-        Self::new_from_index(None, state)
-    }
-
-    /// true for an empty tree
-    pub fn is_empty(&self) -> bool {
-        self.count() == 0
-    }
-
-    /// number of elements in the tree
-    pub fn count(&self) -> u64 {
-        self.root.as_ref().map(|x| x.count()).unwrap_or_default()
-    }
-
-    /// root of a non-empty tree
-    pub fn root(&self) -> Option<&T::Link> {
-        self.root.as_ref().and_then(|index| index.link().as_ref())
     }
 }
